pyproject.toml
[tool.poetry]
name = "daera-clear"
version = "0.1.0"
description = ""
authors = ["ahndong <yachtahn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.5"
annotated-types = "0.5.0"
anyio = "3.7.1"
asyncpg = "0.28.0"
click = "8.1.6"
colorama = "0.4.6"
exceptiongroup = "1.1.2"
fastapi = "0.101.0"
graphql-core = "3.2.3"
greenlet = "2.0.2"
h11 = "0.14.0"
idna = "3.4"
psycopg2 = "2.9.7"
pydantic = "1.10.12"
pydantic_core = "2.4.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
six = "1.16.0"
sniffio = "1.3.0"
SQLAlchemy = "1.4.40"
sqlalchemy2-stubs = "0.0.2a35"
sqlmodel = "0.0.8"
starlette = "0.27.0"
strawberry-graphql = "0.203.0"
typing_extensions = "4.7.1"
uvicorn = "0.23.2"

[tool.poetry.group.dev.dependencies]
Jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
